[IMP] point_of_sale: Handling product attributes
When selecting an attribute value in POS, what we store is the `description`,
which is a string that represents the selected attribute values.
Ex: selecting `Size: M` and `Material: Leather` will result in the description:
`(L, Leather)`.

This choice does not lead to a logical API for dealing with product attibutes.

In this pr, we keep the `description` field but add a new field that stores
the selected `ids` of `"product.template.attribute.value"`.

closes odoo/odoo#126398

Task: 3378533
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>